Refactor: Extract Markdown list and link formatting logic
Moves the automatic list continuation and interactive link insertion features from `CreateVoiceMarkdownFormatting.kt` into new, dedicated extension files: `CreateVoiceListContinuation.kt` and `CreateVoiceMarkdownLinks.kt`. This improves modularity and further reduces the complexity of the main markdown formatting file, continuing the effort to organize `CreateVoiceActivity` related code.
